+++ Runtime/ContinuousDistributions/XorShift32Distribution/Generators/XorShift32GeneratorRanged.cs
@@ -77,6 +77,12 @@ namespace Zor.RandomGenerators.ContinuousDistributions
 		}
 
 		/// <inheritdoc/>
+		public void Forward(int steps)
+		{
+			m_randomEngine.Forward(steps);
+		}
+
+		/// <inheritdoc/>
 		[Pure]
 		public float Generate()
 		{