+++ Runtime/DiscreteDistributions/XorShift128Distribution/Generators/IntXorShift128Generator.cs
@@ -113,6 +113,12 @@ namespace Zor.RandomGenerators.DiscreteDistributions
 		}
 
 		/// <inheritdoc/>
+		public void Forward(int steps)
+		{
+			m_randomEngine.Forward(steps);
+		}
+
+		/// <inheritdoc/>
 		[Pure]
 		public int Generate()
 		{